The Great Merge of 2010
git-tfs-id: [http://localhost:8080/tfs/DefaultCollection/]$/foobar2000/files/plugins.root;C370

+++ VSTiPlayer.h
@@ -11,7 +11,7 @@ struct VstTimeInfo;
 #endif
 
 #ifndef _AUDIO_CHUNK_H_
-typedef double audio_sample;
+typedef float audio_sample;
 #endif
 
 class VSTiPlayer
